+++ test/hexapdf/type/test_object_stream.rb
@@ -104,6 +104,15 @@ describe HexaPDF::Type::ObjectStream do
       assert_equal(0, @obj.value[:First])
       assert_equal("", @obj.stream)
     end
+
+    it "doesn't allow signature dictionaries to be compressed" do
+      @obj.add_object(HexaPDF::Dictionary.new({Type: :Sig}, oid: 1))
+      @obj.add_object(HexaPDF::Dictionary.new({Type: :DocTimeStamp}, oid: 2))
+      @obj.add_object(HexaPDF::Dictionary.new({ByteRange: [], Contents: ''}, oid: 3))
+      @obj.write_objects(@revision)
+      assert_equal(0, @obj.value[:N])
+      assert_equal("", @obj.stream)
+    end
   end
 
   it "fails validation if gen != 0" do