+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [HexaPDF::Object#==] to allow comparison to simple value if not indirect
 * [HexaPDF::Type::AcroForm::Form] to use an empty array as default for the
   /Fields field
+* [HexaPDF::Type::ObjectStream] to not store signature fields in object streams
 
 
 ## 0.19.3 - 2021-12-14